# API/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/collect/workstation": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/collect/specs": {
      "post": {
        "description": "",
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/specs": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/spec/id": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/update/spec": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/delete/spec": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/workstations": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/workstation/id": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/pairs": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/collect/pair": {
      "post": {
        "description": "",
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/modify/pair": {
      "post": {
        "description": "",
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/delete/pair": {
      "post": {
        "description": "",
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/data/monitoring": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}